requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
ddgs==9.5.2
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
duckduckgo-search==8.1.1
ecdsa==0.19.1
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
ffmpy==0.6.1
filelock==3.18.0
fqdn==1.5.1
fsspec==2025.7.0
gradio==5.42.0
gradio-client==1.11.1
griffe==1.11.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
lark==1.2.2
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.12.4
mdurl==0.1.2
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook-shim==0.2.4
numpy==2.3.2
openai==1.99.3
openai-agents==0.2.5
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
primp==0.15.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
psutil==7.0.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==2.0.15
pyyaml==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.8
safehttpx==0.1.6
semantic-version==2.10.0
send2trash==1.8.3
sendgrid==6.12.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
tavily-python==0.7.10
terminado==0.18.1
tiktoken==0.10.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250611
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
